refactor(Car/Cars): 패키지 이동
- test 수정
- 객체가 분리되면서 필요없는 메서드 삭제
