+++ src/main/kotlin/racingcar/domain/Cars.kt
@@ -1,0 +1,9 @@
+package racingcar.domain
+
+class Cars(
+    private val cars: List<Car>
+) {
+    fun getCars() = cars
+
+    fun race() = cars.map { it.moveOrNot() }
+}
